Isolate file drop behavior from the chat input widget
Constraint: Keep dragged-file-path behavior unchanged while reducing future input-widget conflicts
Rejected: Leave DND parsing inside TextInputWidget | it keeps unrelated input features coupled
Confidence: high
Scope-risk: narrow
Directive: Keep future input behaviors in focused support classes when they do not need widget layout ownership
Tested: mvn clean verify
Not-tested: Manual drag/drop in Eclipse UI
